# backend/openapi-gen/src/main/java/se/ifmo/gen/api/DragonsApi.java
/*
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.15.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package se.ifmo.gen.api;

import se.ifmo.gen.model.Coordinates;
import se.ifmo.gen.model.Dragon;
import se.ifmo.gen.model.DragonCave;
import se.ifmo.gen.model.DragonCreate;
import se.ifmo.gen.model.DragonHead;
import se.ifmo.gen.model.GetDragons200Response;
import se.ifmo.gen.model.GetDragonsRequest;
import se.ifmo.gen.model.Location;
import se.ifmo.gen.model.Person;
import se.ifmo.gen.model.ReassignAndDeleteDragonRequest;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.media.ExampleObject;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.lang.Nullable;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", comments = "Generator version: 7.15.0")
@Validated
@Tag(name = "Dragons", description = "Управление драконами")
public interface DragonsApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    public static final String PATH_CREATE_DRAGON = "/dragons";
    /**
     * POST /dragons : Create a dragon
     *
     * @param dragonCreate Dragon to create (required)
     * @return Created (status code 201)
     */
    @Operation(
        operationId = "createDragon",
        summary = "Create a dragon",
        tags = { "Dragons" },
        responses = {
            @ApiResponse(responseCode = "201", description = "Created")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = DragonsApi.PATH_CREATE_DRAGON,
        consumes = { "application/json" }
    )
    
    default ResponseEntity<Void> createDragon(
        @Parameter(name = "DragonCreate", description = "Dragon to create", required = true) @Valid @RequestBody DragonCreate dragonCreate
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    public static final String PATH_DELETE_DRAGON_BY_AGE = "/dragons/ages/{age}";
    /**
     * DELETE /dragons/ages/{age} : Delete a dragon by age
     *
     * @param age Dragon age (required)
     * @return No content (status code 204)
     */
    @Operation(
        operationId = "deleteDragonByAge",
        summary = "Delete a dragon by age",
        tags = { "Dragons" },
        responses = {
            @ApiResponse(responseCode = "204", description = "No content")
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = DragonsApi.PATH_DELETE_DRAGON_BY_AGE
    )
    
    default ResponseEntity<Void> deleteDragonByAge(
        @Parameter(name = "age", description = "Dragon age", required = true, in = ParameterIn.PATH) @PathVariable("age") Integer age
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    public static final String PATH_DELETE_DRAGON_BY_ID = "/dragons/{id}";
    /**
     * DELETE /dragons/{id} : Delete a dragon by ID
     *
     * @param id Dragon ID (required)
     * @return No content (status code 204)
     */
    @Operation(
        operationId = "deleteDragonById",
        summary = "Delete a dragon by ID",
        tags = { "Dragons" },
        responses = {
            @ApiResponse(responseCode = "204", description = "No content")
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = DragonsApi.PATH_DELETE_DRAGON_BY_ID
    )
    
    default ResponseEntity<Void> deleteDragonById(
        @Parameter(name = "id", description = "Dragon ID", required = true, in = ParameterIn.PATH) @PathVariable("id") Integer id
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    public static final String PATH_DELETE_DRAGONS_IN_CAVE_WITH_MAX_DEPTH = "/dragons/caves/delete-by-max-depth";
    /**
     * DELETE /dragons/caves/delete-by-max-depth : Delete all dragons in cave with max depth
     *
     * @return No content (status code 204)
     */
    @Operation(
        operationId = "deleteDragonsInCaveWithMaxDepth",
        summary = "Delete all dragons in cave with max depth",
        tags = { "Dragons" },
        responses = {
            @ApiResponse(responseCode = "204", description = "No content")
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = DragonsApi.PATH_DELETE_DRAGONS_IN_CAVE_WITH_MAX_DEPTH
    )
    
    default ResponseEntity<Void> deleteDragonsInCaveWithMaxDepth(
        
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    public static final String PATH_GET_CAVES = "/dragons/caves";
    /**
     * GET /dragons/caves : Get all caves
     *
     * @return OK (status code 200)
     */
    @Operation(
        operationId = "getCaves",
        summary = "Get all caves",
        tags = { "Dragons" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = DragonCave.class)))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = DragonsApi.PATH_GET_CAVES,
        produces = { "application/json" }
    )
    
    default ResponseEntity<List<DragonCave>> getCaves(
        
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "[ { \"depth\" : 2.302136, \"id\" : 1, \"numberOfTreasures\" : 1 }, { \"depth\" : 2.302136, \"id\" : 1, \"numberOfTreasures\" : 1 } ]";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    public static final String PATH_GET_COORDINATES = "/dragons/coordinates";
    /**
     * GET /dragons/coordinates : Get all coordinates
     *
     * @return OK (status code 200)
     */
    @Operation(
        operationId = "getCoordinates",
        summary = "Get all coordinates",
        tags = { "Dragons" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = Coordinates.class)))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = DragonsApi.PATH_GET_COORDINATES,
        produces = { "application/json" }
    )
    
    default ResponseEntity<List<Coordinates>> getCoordinates(
        
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "[ { \"x\" : 27.410704, \"y\" : 5.962133916683182, \"id\" : 1 }, { \"x\" : 27.410704, \"y\" : 5.962133916683182, \"id\" : 1 } ]";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    public static final String PATH_GET_DRAGON_BY_ID = "/dragons/{id}";
    /**
     * GET /dragons/{id} : Get a dragon by ID
     *
     * @param id Dragon ID (required)
     * @return OK (status code 200)
     */
    @Operation(
        operationId = "getDragonById",
        summary = "Get a dragon by ID",
        tags = { "Dragons" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Dragon.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = DragonsApi.PATH_GET_DRAGON_BY_ID,
        produces = { "application/json" }
    )
    
    default ResponseEntity<Dragon> getDragonById(
        @Parameter(name = "id", description = "Dragon ID", required = true, in = ParameterIn.PATH) @PathVariable("id") Integer id
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"head\" : { \"toothCount\" : 2.027123, \"id\" : 1 }, \"character\" : \"\", \"cave\" : { \"depth\" : 2.302136, \"id\" : 1, \"numberOfTreasures\" : 1 }, \"color\" : \"GREEN\", \"name\" : \"name\", \"coordinates\" : { \"x\" : 27.410704, \"y\" : 5.962133916683182, \"id\" : 1 }, \"id\" : 1, \"killer\" : \"\", \"creationDate\" : \"2000-01-23T04:56:07.000+00:00\", \"type\" : \"WATER\", \"age\" : 1 }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    public static final String PATH_GET_DRAGONS = "/dragons/filter";
    /**
     * POST /dragons/filter : Get all dragons
     *
     * @param getDragonsRequest Filters (required)
     * @return OK (status code 200)
     */
    @Operation(
        operationId = "getDragons",
        summary = "Get all dragons",
        tags = { "Dragons" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = GetDragons200Response.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = DragonsApi.PATH_GET_DRAGONS,
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    default ResponseEntity<GetDragons200Response> getDragons(
        @Parameter(name = "GetDragonsRequest", description = "Filters", required = true) @Valid @RequestBody GetDragonsRequest getDragonsRequest
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"total_page_count\" : 0, \"total_dragons_count\" : 6, \"dragons\" : [ { \"head\" : { \"toothCount\" : 2.027123, \"id\" : 1 }, \"character\" : \"\", \"cave\" : { \"depth\" : 2.302136, \"id\" : 1, \"numberOfTreasures\" : 1 }, \"color\" : \"GREEN\", \"name\" : \"name\", \"coordinates\" : { \"x\" : 27.410704, \"y\" : 5.962133916683182, \"id\" : 1 }, \"id\" : 1, \"killer\" : \"\", \"creationDate\" : \"2000-01-23T04:56:07.000+00:00\", \"type\" : \"WATER\", \"age\" : 1 }, { \"head\" : { \"toothCount\" : 2.027123, \"id\" : 1 }, \"character\" : \"\", \"cave\" : { \"depth\" : 2.302136, \"id\" : 1, \"numberOfTreasures\" : 1 }, \"color\" : \"GREEN\", \"name\" : \"name\", \"coordinates\" : { \"x\" : 27.410704, \"y\" : 5.962133916683182, \"id\" : 1 }, \"id\" : 1, \"killer\" : \"\", \"creationDate\" : \"2000-01-23T04:56:07.000+00:00\", \"type\" : \"WATER\", \"age\" : 1 } ] }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    public static final String PATH_GET_DRAGONS_AGE_SUM = "/dragons/ages/sum";
    /**
     * GET /dragons/ages/sum : Get sum of all dragons ages
     *
     * @return OK (status code 200)
     */
    @Operation(
        operationId = "getDragonsAgeSum",
        summary = "Get sum of all dragons ages",
        tags = { "Dragons" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Integer.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = DragonsApi.PATH_GET_DRAGONS_AGE_SUM,
        produces = { "application/json" }
    )
    
    default ResponseEntity<Integer> getDragonsAgeSum(
        
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    public static final String PATH_GET_HEADS = "/dragons/heads";
    /**
     * GET /dragons/heads : Get all heads
     *
     * @return OK (status code 200)
     */
    @Operation(
        operationId = "getHeads",
        summary = "Get all heads",
        tags = { "Dragons" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = DragonHead.class)))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = DragonsApi.PATH_GET_HEADS,
        produces = { "application/json" }
    )
    
    default ResponseEntity<List<DragonHead>> getHeads(
        
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "[ { \"toothCount\" : 2.027123, \"id\" : 1 }, { \"toothCount\" : 2.027123, \"id\" : 1 } ]";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    public static final String PATH_GET_LOCATIONS = "/dragons/locations";
    /**
     * GET /dragons/locations : Get all locations
     *
     * @return OK (status code 200)
     */
    @Operation(
        operationId = "getLocations",
        summary = "Get all locations",
        tags = { "Dragons" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = Location.class)))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = DragonsApi.PATH_GET_LOCATIONS,
        produces = { "application/json" }
    )
    
    default ResponseEntity<List<Location>> getLocations(
        
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "[ { \"x\" : 6.027456183070403, \"name\" : \"name\", \"y\" : 1, \"id\" : 1 }, { \"x\" : 6.027456183070403, \"name\" : \"name\", \"y\" : 1, \"id\" : 1 } ]";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    public static final String PATH_GET_MAX_TYPE_DRAGON = "/dragons/types/max";
    /**
     * GET /dragons/types/max : Get dragon with max type
     *
     * @return OK (status code 200)
     */
    @Operation(
        operationId = "getMaxTypeDragon",
        summary = "Get dragon with max type",
        tags = { "Dragons" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Dragon.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = DragonsApi.PATH_GET_MAX_TYPE_DRAGON,
        produces = { "application/json" }
    )
    
    default ResponseEntity<Dragon> getMaxTypeDragon(
        
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"head\" : { \"toothCount\" : 2.027123, \"id\" : 1 }, \"character\" : \"\", \"cave\" : { \"depth\" : 2.302136, \"id\" : 1, \"numberOfTreasures\" : 1 }, \"color\" : \"GREEN\", \"name\" : \"name\", \"coordinates\" : { \"x\" : 27.410704, \"y\" : 5.962133916683182, \"id\" : 1 }, \"id\" : 1, \"killer\" : \"\", \"creationDate\" : \"2000-01-23T04:56:07.000+00:00\", \"type\" : \"WATER\", \"age\" : 1 }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    public static final String PATH_GET_PERSONS = "/dragons/persons";
    /**
     * GET /dragons/persons : Get all persons
     *
     * @return OK (status code 200)
     */
    @Operation(
        operationId = "getPersons",
        summary = "Get all persons",
        tags = { "Dragons" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = Person.class)))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = DragonsApi.PATH_GET_PERSONS,
        produces = { "application/json" }
    )
    
    default ResponseEntity<List<Person>> getPersons(
        
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "[ { \"eyeColor\" : \"GREEN\", \"name\" : \"name\", \"weight\" : 0.14658229805029452, \"location\" : \"\", \"id\" : 1, \"hairColor\" : \"\", \"passportID\" : \"passportID\", \"height\" : 0.6027466183070403 }, { \"eyeColor\" : \"GREEN\", \"name\" : \"name\", \"weight\" : 0.14658229805029452, \"location\" : \"\", \"id\" : 1, \"hairColor\" : \"\", \"passportID\" : \"passportID\", \"height\" : 0.6027466183070403 } ]";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    public static final String PATH_KILL_DRAGON = "/dragons/persons/{person-id}/kill/{dragon-id}";
    /**
     * POST /dragons/persons/{person-id}/kill/{dragon-id} : Kill dragon
     *
     * @param personId Killer person ID (required)
     * @param dragonId Dragon ID (required)
     * @return OK (status code 200)
     */
    @Operation(
        operationId = "killDragon",
        summary = "Kill dragon",
        tags = { "Dragons" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = DragonsApi.PATH_KILL_DRAGON
    )
    
    default ResponseEntity<Void> killDragon(
        @Parameter(name = "person-id", description = "Killer person ID", required = true, in = ParameterIn.PATH) @PathVariable("person-id") Integer personId,
        @Parameter(name = "dragon-id", description = "Dragon ID", required = true, in = ParameterIn.PATH) @PathVariable("dragon-id") Integer dragonId
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    public static final String PATH_REASSIGN_AND_DELETE_DRAGON = "/dragons/{id}/reassign-and-delete";
    /**
     * POST /dragons/{id}/reassign-and-delete : Reassign dependencies and delete a dragon
     *
     * @param id ID of the dragon to delete (required)
     * @param reassignAndDeleteDragonRequest IDs of dragon to reassign dependencies to. (required)
     * @return Successfully deleted (status code 204)
     *         or Dragon not found (status code 404)
     *         or Bad request (e.g., new owner ID is invalid) (status code 400)
     */
    @Operation(
        operationId = "reassignAndDeleteDragon",
        summary = "Reassign dependencies and delete a dragon",
        tags = { "Dragons" },
        responses = {
            @ApiResponse(responseCode = "204", description = "Successfully deleted"),
            @ApiResponse(responseCode = "404", description = "Dragon not found"),
            @ApiResponse(responseCode = "400", description = "Bad request (e.g., new owner ID is invalid)")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = DragonsApi.PATH_REASSIGN_AND_DELETE_DRAGON,
        consumes = { "application/json" }
    )
    
    default ResponseEntity<Void> reassignAndDeleteDragon(
        @Parameter(name = "id", description = "ID of the dragon to delete", required = true, in = ParameterIn.PATH) @PathVariable("id") Integer id,
        @Parameter(name = "ReassignAndDeleteDragonRequest", description = "IDs of dragon to reassign dependencies to.", required = true) @Valid @RequestBody ReassignAndDeleteDragonRequest reassignAndDeleteDragonRequest
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    public static final String PATH_UPDATE_DRAGON = "/dragons";
    /**
     * PUT /dragons : Update a dragon
     *
     * @param dragon Dragon to update (required)
     * @return OK (status code 200)
     */
    @Operation(
        operationId = "updateDragon",
        summary = "Update a dragon",
        tags = { "Dragons" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK")
        }
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = DragonsApi.PATH_UPDATE_DRAGON,
        consumes = { "application/json" }
    )
    
    default ResponseEntity<Void> updateDragon(
        @Parameter(name = "Dragon", description = "Dragon to update", required = true) @Valid @RequestBody Dragon dragon
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}